requirements.txt
langchain==0.1.9
packaging
requests
loguru
PyYAML
python-telegram-bot==20.8
chalice
langchain-openai